+++ src/city/tests/mock/MockRestaurantZhangCustomer.java
@@ -18,11 +18,11 @@ public class MockRestaurantZhangCustomer extends MockRole implements RestaurantZ
 		log.add(new LoggedEvent("Received check from waiter."));
 	}
 	
-	public void msgHereIsChange(double change) {
+	public void msgHereIsChange(int change) {
 		log.add(new LoggedEvent("Got change " + change + " from cashier."));
 	}
 	
-	public void msgPayLater(double tab) {
+	public void msgPayLater(int tab) {
 		log.add(new LoggedEvent("Got tab " + tab + " from cashier."));
 	}
